at great pain and struggle to myself, I have set up a way for my frontend to communicate to my backend. This will allow my api keys to remain safe. now I have to fix all the bugs that are showing up in the frontend.
