+++ .gitignore
@@ -23,4 +23,4 @@ yarn-debug.log*
 yarn-error.log*
 
 package-lock.json
-src/config.js
+.env
